♻️ refactor:  불필요한 코드 제거 / 상품 리스트 limit 제한 해제
